requirements.txt
# -----------------------------
# UI
# -----------------------------
streamlit>=1.31.0

# -----------------------------
# Config / env
# -----------------------------
python-dotenv>=1.0.0

# -----------------------------
# OpenAI
# -----------------------------
openai>=1.12.0

# -----------------------------
# Vector search / data
# -----------------------------
faiss-cpu>=1.7.4
numpy>=1.26.0
pandas>=2.1.0

# -----------------------------
# PDF extraction (digital PDFs)
# -----------------------------
pdfplumber>=0.10.3
PyPDF2>=3.0.1

# -----------------------------
# OCR fallback (scanned PDFs)
# -----------------------------
pymupdf>=1.23.0
pillow>=10.2.0
pytesseract>=0.3.10

# -----------------------------
# Optional helpful utilities
# -----------------------------
tqdm>=4.66.0
